Do not use resource filters

+++ extensions-core/druid-basic-security/src/main/java/org/apache/druid/security/basic/authorization/endpoint/BasicAuthorizerResource.java
@@ -23,9 +23,9 @@ import com.fasterxml.jackson.jaxrs.smile.SmileMediaTypes;
 import com.google.inject.Inject;
 import com.sun.jersey.spi.container.ResourceFilters;
 import org.apache.druid.guice.LazySingleton;
-import org.apache.druid.indexing.overlord.http.security.AuthorizerResourceFilter;
 import org.apache.druid.security.basic.BasicSecurityResourceFilter;
 import org.apache.druid.security.basic.authorization.entity.BasicAuthorizerGroupMapping;
+import org.apache.druid.server.security.AuthorizerValidator;
 import org.apache.druid.server.security.ResourceAction;
 
 import javax.servlet.http.HttpServletRequest;
@@ -47,13 +47,16 @@ import java.util.List;
 public class BasicAuthorizerResource
 {
   private final BasicAuthorizerResourceHandler resourceHandler;
+  private final AuthorizerValidator authorizerValidator;
 
   @Inject
   public BasicAuthorizerResource(
-      BasicAuthorizerResourceHandler resourceHandler
+      BasicAuthorizerResourceHandler resourceHandler,
+      AuthorizerValidator authorizerValidator
   )
   {
     this.resourceHandler = resourceHandler;
+    this.authorizerValidator = authorizerValidator;
   }
 
   /**
@@ -101,12 +104,13 @@ public class BasicAuthorizerResource
   @Path("/db/{authorizerName}/users")
   @Produces(MediaType.APPLICATION_JSON)
   @Consumes(MediaType.APPLICATION_JSON)
-  @ResourceFilters({BasicSecurityResourceFilter.class, AuthorizerResourceFilter.class})
+  @ResourceFilters(BasicSecurityResourceFilter.class)
   public Response getAllUsers(
       @Context HttpServletRequest req,
       @PathParam("authorizerName") final String authorizerName
   )
   {
+    authorizerValidator.validateAuthorizerName(authorizerName);
     return resourceHandler.getAllUsers(authorizerName);
   }
 
@@ -119,12 +123,13 @@ public class BasicAuthorizerResource
   @Path("/db/{authorizerName}/groupMappings")
   @Produces(MediaType.APPLICATION_JSON)
   @Consumes(MediaType.APPLICATION_JSON)
-  @ResourceFilters({BasicSecurityResourceFilter.class, AuthorizerResourceFilter.class})
+  @ResourceFilters(BasicSecurityResourceFilter.class)
   public Response getAllGroupMappings(
       @Context HttpServletRequest req,
       @PathParam("authorizerName") final String authorizerName
   )
   {
+    authorizerValidator.validateAuthorizerName(authorizerName);
     return resourceHandler.getAllGroupMappings(authorizerName);
   }
 
@@ -138,7 +143,7 @@ public class BasicAuthorizerResource
   @Path("/db/{authorizerName}/users/{userName}")
   @Produces(MediaType.APPLICATION_JSON)
   @Consumes(MediaType.APPLICATION_JSON)
-  @ResourceFilters({BasicSecurityResourceFilter.class, AuthorizerResourceFilter.class})
+  @ResourceFilters(BasicSecurityResourceFilter.class)
   public Response getUser(
       @Context HttpServletRequest req,
       @PathParam("authorizerName") final String authorizerName,
@@ -147,6 +152,7 @@ public class BasicAuthorizerResource
       @QueryParam("simplifyPermissions") String simplifyPermissions
   )
   {
+    authorizerValidator.validateAuthorizerName(authorizerName);
     return resourceHandler.getUser(authorizerName, userName, full != null, simplifyPermissions != null);
   }
 
@@ -160,7 +166,7 @@ public class BasicAuthorizerResource
   @Path("/db/{authorizerName}/groupMappings/{groupMappingName}")
   @Produces(MediaType.APPLICATION_JSON)
   @Consumes(MediaType.APPLICATION_JSON)
-  @ResourceFilters({BasicSecurityResourceFilter.class, AuthorizerResourceFilter.class})
+  @ResourceFilters(BasicSecurityResourceFilter.class)
   public Response getGroupMapping(
       @Context HttpServletRequest req,
       @PathParam("authorizerName") final String authorizerName,
@@ -168,6 +174,7 @@ public class BasicAuthorizerResource
       @QueryParam("full") String full
   )
   {
+    authorizerValidator.validateAuthorizerName(authorizerName);
     return resourceHandler.getGroupMapping(authorizerName, groupMappingName, full != null);
   }
 
@@ -183,13 +190,14 @@ public class BasicAuthorizerResource
   @Path("/db/{authorizerName}/users/{userName}")
   @Produces(MediaType.APPLICATION_JSON)
   @Consumes(MediaType.APPLICATION_JSON)
-  @ResourceFilters({BasicSecurityResourceFilter.class, AuthorizerResourceFilter.class})
+  @ResourceFilters(BasicSecurityResourceFilter.class)
   public Response createUser(
       @Context HttpServletRequest req,
       @PathParam("authorizerName") final String authorizerName,
       @PathParam("userName") String userName
   )
   {
+    authorizerValidator.validateAuthorizerName(authorizerName);
     return resourceHandler.createUser(authorizerName, userName);
   }
 
@@ -205,13 +213,14 @@ public class BasicAuthorizerResource
   @Path("/db/{authorizerName}/users/{userName}")
   @Produces(MediaType.APPLICATION_JSON)
   @Consumes(MediaType.APPLICATION_JSON)
-  @ResourceFilters({BasicSecurityResourceFilter.class, AuthorizerResourceFilter.class})
+  @ResourceFilters(BasicSecurityResourceFilter.class)
   public Response deleteUser(
       @Context HttpServletRequest req,
       @PathParam("authorizerName") final String authorizerName,
       @PathParam("userName") String userName
   )
   {
+    authorizerValidator.validateAuthorizerName(authorizerName);
     return resourceHandler.deleteUser(authorizerName, userName);
   }
 
@@ -227,7 +236,7 @@ public class BasicAuthorizerResource
   @Path("/db/{authorizerName}/groupMappings/{groupMappingName}")
   @Produces(MediaType.APPLICATION_JSON)
   @Consumes(MediaType.APPLICATION_JSON)
-  @ResourceFilters({BasicSecurityResourceFilter.class, AuthorizerResourceFilter.class})
+  @ResourceFilters(BasicSecurityResourceFilter.class)
   public Response createGroupMapping(
       @Context HttpServletRequest req,
       @PathParam("authorizerName") final String authorizerName,
@@ -235,6 +244,7 @@ public class BasicAuthorizerResource
       BasicAuthorizerGroupMapping groupMapping
   )
   {
+    authorizerValidator.validateAuthorizerName(authorizerName);
     return resourceHandler.createGroupMapping(
         authorizerName,
         new BasicAuthorizerGroupMapping(groupMappingName, groupMapping.getGroupPattern(), groupMapping.getRoles())
@@ -253,13 +263,14 @@ public class BasicAuthorizerResource
   @Path("/db/{authorizerName}/groupMappings/{groupMappingName}")
   @Produces(MediaType.APPLICATION_JSON)
   @Consumes(MediaType.APPLICATION_JSON)
-  @ResourceFilters({BasicSecurityResourceFilter.class, AuthorizerResourceFilter.class})
+  @ResourceFilters(BasicSecurityResourceFilter.class)
   public Response deleteGroupMapping(
       @Context HttpServletRequest req,
       @PathParam("authorizerName") final String authorizerName,
       @PathParam("groupMappingName") String groupMappingName
   )
   {
+    authorizerValidator.validateAuthorizerName(authorizerName);
     return resourceHandler.deleteGroupMapping(authorizerName, groupMappingName);
   }
 
@@ -272,12 +283,13 @@ public class BasicAuthorizerResource
   @Path("/db/{authorizerName}/roles")
   @Produces(MediaType.APPLICATION_JSON)
   @Consumes(MediaType.APPLICATION_JSON)
-  @ResourceFilters({BasicSecurityResourceFilter.class, AuthorizerResourceFilter.class})
+  @ResourceFilters(BasicSecurityResourceFilter.class)
   public Response getAllRoles(
       @Context HttpServletRequest req,
       @PathParam("authorizerName") final String authorizerName
   )
   {
+    authorizerValidator.validateAuthorizerName(authorizerName);
     return resourceHandler.getAllRoles(authorizerName);
   }
 
@@ -293,7 +305,7 @@ public class BasicAuthorizerResource
   @Path("/db/{authorizerName}/roles/{roleName}")
   @Produces(MediaType.APPLICATION_JSON)
   @Consumes(MediaType.APPLICATION_JSON)
-  @ResourceFilters({BasicSecurityResourceFilter.class, AuthorizerResourceFilter.class})
+  @ResourceFilters(BasicSecurityResourceFilter.class)
   public Response getRole(
       @Context HttpServletRequest req,
       @PathParam("authorizerName") final String authorizerName,
@@ -302,6 +314,7 @@ public class BasicAuthorizerResource
       @QueryParam("simplifyPermissions") String simplifyPermissions
   )
   {
+    authorizerValidator.validateAuthorizerName(authorizerName);
     return resourceHandler.getRole(authorizerName, roleName, full != null, simplifyPermissions != null);
   }
 
@@ -317,13 +330,14 @@ public class BasicAuthorizerResource
   @Path("/db/{authorizerName}/roles/{roleName}")
   @Produces(MediaType.APPLICATION_JSON)
   @Consumes(MediaType.APPLICATION_JSON)
-  @ResourceFilters({BasicSecurityResourceFilter.class, AuthorizerResourceFilter.class})
+  @ResourceFilters(BasicSecurityResourceFilter.class)
   public Response createRole(
       @Context HttpServletRequest req,
       @PathParam("authorizerName") final String authorizerName,
       @PathParam("roleName") final String roleName
   )
   {
+    authorizerValidator.validateAuthorizerName(authorizerName);
     return resourceHandler.createRole(authorizerName, roleName);
   }
 
@@ -339,13 +353,14 @@ public class BasicAuthorizerResource
   @Path("/db/{authorizerName}/roles/{roleName}")
   @Produces(MediaType.APPLICATION_JSON)
   @Consumes(MediaType.APPLICATION_JSON)
-  @ResourceFilters({BasicSecurityResourceFilter.class, AuthorizerResourceFilter.class})
+  @ResourceFilters(BasicSecurityResourceFilter.class)
   public Response deleteRole(
       @Context HttpServletRequest req,
       @PathParam("authorizerName") final String authorizerName,
       @PathParam("roleName") String roleName
   )
   {
+    authorizerValidator.validateAuthorizerName(authorizerName);
     return resourceHandler.deleteRole(authorizerName, roleName);
   }
 
@@ -362,7 +377,7 @@ public class BasicAuthorizerResource
   @Path("/db/{authorizerName}/users/{userName}/roles/{roleName}")
   @Produces(MediaType.APPLICATION_JSON)
   @Consumes(MediaType.APPLICATION_JSON)
-  @ResourceFilters({BasicSecurityResourceFilter.class, AuthorizerResourceFilter.class})
+  @ResourceFilters(BasicSecurityResourceFilter.class)
   public Response assignRoleToUser(
       @Context HttpServletRequest req,
       @PathParam("authorizerName") final String authorizerName,
@@ -370,6 +385,7 @@ public class BasicAuthorizerResource
       @PathParam("roleName") String roleName
   )
   {
+    authorizerValidator.validateAuthorizerName(authorizerName);
     return resourceHandler.assignRoleToUser(authorizerName, userName, roleName);
   }
 
@@ -386,7 +402,7 @@ public class BasicAuthorizerResource
   @Path("/db/{authorizerName}/users/{userName}/roles/{roleName}")
   @Produces(MediaType.APPLICATION_JSON)
   @Consumes(MediaType.APPLICATION_JSON)
-  @ResourceFilters({BasicSecurityResourceFilter.class, AuthorizerResourceFilter.class})
+  @ResourceFilters(BasicSecurityResourceFilter.class)
   public Response unassignRoleFromUser(
       @Context HttpServletRequest req,
       @PathParam("authorizerName") final String authorizerName,
@@ -394,6 +410,7 @@ public class BasicAuthorizerResource
       @PathParam("roleName") String roleName
   )
   {
+    authorizerValidator.validateAuthorizerName(authorizerName);
     return resourceHandler.unassignRoleFromUser(authorizerName, userName, roleName);
   }
 
@@ -410,7 +427,7 @@ public class BasicAuthorizerResource
   @Path("/db/{authorizerName}/groupMappings/{groupMappingName}/roles/{roleName}")
   @Produces(MediaType.APPLICATION_JSON)
   @Consumes(MediaType.APPLICATION_JSON)
-  @ResourceFilters({BasicSecurityResourceFilter.class, AuthorizerResourceFilter.class})
+  @ResourceFilters(BasicSecurityResourceFilter.class)
   public Response assignRoleToGroupMapping(
       @Context HttpServletRequest req,
       @PathParam("authorizerName") final String authorizerName,
@@ -418,6 +435,7 @@ public class BasicAuthorizerResource
       @PathParam("roleName") String roleName
   )
   {
+    authorizerValidator.validateAuthorizerName(authorizerName);
     return resourceHandler.assignRoleToGroupMapping(authorizerName, groupMappingName, roleName);
   }
 
@@ -434,7 +452,7 @@ public class BasicAuthorizerResource
   @Path("/db/{authorizerName}/groupMappings/{groupMappingName}/roles/{roleName}")
   @Produces(MediaType.APPLICATION_JSON)
   @Consumes(MediaType.APPLICATION_JSON)
-  @ResourceFilters({BasicSecurityResourceFilter.class, AuthorizerResourceFilter.class})
+  @ResourceFilters(BasicSecurityResourceFilter.class)
   public Response unassignRoleFromGroupMapping(
       @Context HttpServletRequest req,
       @PathParam("authorizerName") final String authorizerName,
@@ -442,6 +460,7 @@ public class BasicAuthorizerResource
       @PathParam("roleName") String roleName
   )
   {
+    authorizerValidator.validateAuthorizerName(authorizerName);
     return resourceHandler.unassignRoleFromGroupMapping(authorizerName, groupMappingName, roleName);
   }
 
@@ -458,7 +477,7 @@ public class BasicAuthorizerResource
   @Path("/db/{authorizerName}/roles/{roleName}/permissions")
   @Produces(MediaType.APPLICATION_JSON)
   @Consumes(MediaType.APPLICATION_JSON)
-  @ResourceFilters({BasicSecurityResourceFilter.class, AuthorizerResourceFilter.class})
+  @ResourceFilters(BasicSecurityResourceFilter.class)
   public Response setRolePermissions(
       @Context HttpServletRequest req,
       @PathParam("authorizerName") final String authorizerName,
@@ -466,6 +485,7 @@ public class BasicAuthorizerResource
       List<ResourceAction> permissions
   )
   {
+    authorizerValidator.validateAuthorizerName(authorizerName);
     return resourceHandler.setRolePermissions(authorizerName, roleName, permissions);
   }
 
@@ -488,6 +508,7 @@ public class BasicAuthorizerResource
       @PathParam("roleName") String roleName
   )
   {
+    authorizerValidator.validateAuthorizerName(authorizerName);
     return resourceHandler.getRolePermissions(authorizerName, roleName);
   }
 
@@ -506,6 +527,7 @@ public class BasicAuthorizerResource
       @PathParam("authorizerName") final String authorizerName
   )
   {
+    authorizerValidator.validateAuthorizerName(authorizerName);
     return resourceHandler.getCachedUserMaps(authorizerName);
   }
 
@@ -524,6 +546,7 @@ public class BasicAuthorizerResource
       @PathParam("authorizerName") final String authorizerName
   )
   {
+    authorizerValidator.validateAuthorizerName(authorizerName);
     return resourceHandler.getCachedGroupMappingMaps(authorizerName);
   }
 
@@ -545,6 +568,7 @@ public class BasicAuthorizerResource
       byte[] serializedUserAndRoleMap
   )
   {
+    authorizerValidator.validateAuthorizerName(authorizerName);
     return resourceHandler.authorizerUserUpdateListener(authorizerName, serializedUserAndRoleMap);
   }
 
@@ -562,6 +586,7 @@ public class BasicAuthorizerResource
       byte[] serializedUserAndRoleMap
   )
   {
+    authorizerValidator.validateAuthorizerName(authorizerName);
     return resourceHandler.authorizerUserUpdateListener(authorizerName, serializedUserAndRoleMap);
   }
 
@@ -579,6 +604,7 @@ public class BasicAuthorizerResource
       byte[] serializedGroupMappingAndRoleMap
   )
   {
+    authorizerValidator.validateAuthorizerName(authorizerName);
     return resourceHandler.authorizerGroupMappingUpdateListener(authorizerName, serializedGroupMappingAndRoleMap);
   }
 }
